don't have dedicated source files for just canRemove
