--- overlord/snapstate/can_remove.go
@@ -1,42 +1,0 @@
-// -*- Mode: Go; indent-tabs-mode: t -*-
-
-/*
- * Copyright (C) 2014-2016 Canonical Ltd
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as
- * published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package snapstate
-
-import (
-	"github.com/snapcore/snapd/snap"
-)
-
-// canRemove verifies that a snap can be removed.
-func canRemove(s *snap.Info, active bool) bool {
-	// Gadget snaps should not be removed as they are a key
-	// building block for Gadgets. Pruning non active ones
-	// is acceptable.
-	if s.Type == snap.TypeGadget && active {
-		return false
-	}
-
-	// You never want to remove an active kernel or OS
-	if (s.Type == snap.TypeKernel || s.Type == snap.TypeOS) && active {
-		return false
-	}
-	// TODO: on classic likely let remove core even if active if it's only snap left.
-
-	return true
-}
